Comment out parentheses enclosing snippet
Commented out the snippet that matches anything and encloses it in
parentheses if it is visually selected. This snippet caused it so that
other snippets would not work when jumping between $1, $2, ... This is
actually due to the fact that I have <tab> as the snippet expanding key
and the "jump to next part of snippet" key. I'll have to think of some
way to get around this, or remap the keys mentioned.