Remove and adjust snippets
Removed some snippets for all file types that would clash with other
snippets. These were rarely used anyway.

Fixed bugs in Julia snippets which allowed all snippets to auto-expand,
when it should not be this way. Instead, I had meant each snippet to
take into account if it was in another word, but had used the
auto-expand feature instead.